update base image + pkg installs + maintainer
The previous version of the Dockerfile would succeed in building for
some reason. This commit fixes that issue and in general cleans up the
dockerfile. Most packages are installed from conda, and those that are
not installable from conda are installed with pip. The base image is
changed to miniconda3 to make the image smaller. Conda takes care of all
of pytorch's dependencies.

I (@kaczmarj) was able to use this Dockerfile to run inference on a
whole slide image. I did this on a CPU and have not tested on GPU.